RP8: NewRenderPipeline Research WIP 3

+++ Code/Core/typedefs.h
@@ -6,11 +6,11 @@
 typedef signed long BOOL32;
 typedef unsigned char U8;
 typedef unsigned short U16;
-typedef unsigned long U32;
+typedef unsigned int U32;
 typedef char C8;
 typedef signed char S8;
 typedef signed short S16;
-typedef signed long S32;
+typedef signed int S32;
 typedef unsigned __int64 U64;
 typedef signed __int64 S64;
 typedef float SINGLE;
